pyproject.toml
[tool.poetry]
name = "rem3"
version = "0.1.0"
description = ""
authors = []
packages = [
  {include = "./src"},
]

[tool.poetry.dependencies]
python = "^3.11.9"
"discord.py" = "^2.3.2"
"psycopg2" = "2.9.8"
pytz = "2024.1"
"python-dotenv" = "*"


[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"